Update to create Chessboard Squares based on manual User Input with chessboard corners
This file is a fallback, if our automated solution does not manage to detect the chessboard squares. Here the user is supposed to mark the chessboard corners. Afterwards through transformation it detects the squares and then through a further function allocates the coordinates to the squares as tuples with lists.